refactor all controller-service and add edit, active-inactive, delete post

+++ src/main/java/com/example/ThreadHub/service/impl/CommunityMemberServiceImpl.java
@@ -4,8 +4,10 @@ import com.example.ThreadHub.dto.response.CommunityMemberResponse;
 import com.example.ThreadHub.entity.Account;
 import com.example.ThreadHub.entity.Community;
 import com.example.ThreadHub.entity.CommunityMember;
+import com.example.ThreadHub.entity.enums.CommunityStatus;
 import com.example.ThreadHub.entity.enums.MemberRole;
 import com.example.ThreadHub.entity.enums.MemberStatus;
+import com.example.ThreadHub.exception.NotFoundException;
 import com.example.ThreadHub.repository.AccountRepository;
 import com.example.ThreadHub.repository.CommunityMemberRepository;
 import com.example.ThreadHub.repository.CommunityRepository;
@@ -15,6 +17,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -42,6 +45,11 @@ public class CommunityMemberServiceImpl implements CommunityMemberService {
         return dto;
     }
 
+    public boolean isMemberExist(Long communityMemberId) {
+        CommunityMember communityMember = communityMemberRepository.findById(communityMemberId).orElse(null);
+        return communityMember != null;
+    }
+
     @Override
     public Page<CommunityMemberResponse> getCommunityMembers(int page, int size, String sortBy, Long CommunityId) {
         Pageable pageable = PageRequest.of(page, size, Sort.by(sortBy).descending());
@@ -56,35 +64,23 @@ public class CommunityMemberServiceImpl implements CommunityMemberService {
         return communityMemberRepository.findByCommunityIdAndAccountId(CommunityId, accountId);
     }
 
-    @Override
-    public boolean isOwner(Long communityMemberId) {
+
+    public CommunityMemberResponse assignModerator(Long communityMemberId) {
+        if(!isMemberExist(communityMemberId)){
+            throw new NotFoundException("cannot find member");
+        }
         CommunityMember communityMember = communityMemberRepository.findById(communityMemberId).orElseThrow();
-        return communityMember.getMemberRole().equals(MemberRole.OWNER);
-    }
-
-    @Override
-    public boolean isModerator(Long communityMemberId) {
-        CommunityMember communityMember = communityMemberRepository.findById(communityMemberId).orElseThrow();
-        return communityMember.getMemberRole().equals(MemberRole.MODERATOR);
-    }
-
-    @Override
-    public boolean isMemberExist(Long communityMemberId) {
-        CommunityMember communityMember = communityMemberRepository.findById(communityMemberId).orElse(null);
-        return communityMember != null;
-    }
-
-    @Override
-    public CommunityMemberResponse assignModerator(Long CommunityMemberId) {
-        CommunityMember communityMember = communityMemberRepository.findById(CommunityMemberId).orElseThrow();
         communityMember.setMemberRole(MemberRole.MODERATOR);
         communityMemberRepository.save(communityMember);
         return mapToDto(communityMember);
     }
 
     @Override
-    public CommunityMemberResponse removeModerator(Long CommunityMemberId) {
-        CommunityMember communityMember = communityMemberRepository.findById(CommunityMemberId).orElseThrow();
+    public CommunityMemberResponse removeModerator(Long communityMemberId) {
+        if(!isMemberExist(communityMemberId)){
+            throw new NotFoundException("cannot find member");
+        }
+        CommunityMember communityMember = communityMemberRepository.findById(communityMemberId).orElseThrow();
         communityMember.setMemberRole(MemberRole.MEMBER);
         communityMemberRepository.save(communityMember);
 
@@ -93,6 +89,9 @@ public class CommunityMemberServiceImpl implements CommunityMemberService {
 
     @Override
     public CommunityMemberResponse transferOwner(Long actorCommunityMemberId, Long newOwnerCommunityMemberId) {
+        if(!isMemberExist(newOwnerCommunityMemberId)){
+            throw new NotFoundException("cannot find member");
+        }
         CommunityMember newOwnerCommunityMember = communityMemberRepository.findById(newOwnerCommunityMemberId).orElseThrow();
         newOwnerCommunityMember.setMemberRole(MemberRole.OWNER);
         communityMemberRepository.save(newOwnerCommunityMember);
@@ -106,9 +105,23 @@ public class CommunityMemberServiceImpl implements CommunityMemberService {
 
     @Override
     public void joinCommunity(Long communityId, Long accountId) {
+        CommunityMember actorCommunityMember = communityMemberRepository.findByCommunityIdAndAccountId(communityId, accountId);
+        Community community = communityRepository.findById(communityId).orElse(null);
+        Account account = accountRepository.findById(accountId).orElse(null);
+
+        if (community == null) {
+            throw new RuntimeException("Community not found");
+        }
+
+        if (community.getCommunityStatus() == CommunityStatus.INACTIVE) {
+            throw new RuntimeException("Community is inactive");
+        }
+
+        if (actorCommunityMember != null) {
+            throw new RuntimeException("You are already a member of this community");
+        }
+
         CommunityMember communityMember = new CommunityMember();
-        Account account = accountRepository.findById(accountId).orElseThrow();
-        Community community = communityRepository.findById(communityId).orElseThrow();
         communityMember.setCommunity(community);
         communityMember.setAccount(account);
         communityMember.setMemberRole(MemberRole.MEMBER);
@@ -118,7 +131,17 @@ public class CommunityMemberServiceImpl implements CommunityMemberService {
 
     @Override
     @Transactional
-    public void leaveCommunity(CommunityMember communityMember) {
+    public void leaveCommunity(Long communityId, Long accountId) {
+        CommunityMember communityMember = communityMemberRepository.findByCommunityIdAndAccountId(communityId, accountId);
+
+        if (communityMember == null) {
+            throw new RuntimeException("You are not a member of this community");
+        }
+
+        if (communityMember.getMemberRole() == MemberRole.OWNER) {
+            throw new RuntimeException("You need to transfer Owner before leaving community");
+        }
+
         communityMemberRepository.delete(communityMember);
         communityMemberRepository.flush();
     }
